Stones, Invisible Walls, Rebaked Nav
Updated the overall level design with stones (Possibly replace some with trees in future). Invisible walls were added to prevent player from jumping out of the borders utilizing the stones.

Rebaked the nav to include all walls and the invisible walls. Stones were not baked.

Added a door for next update as a way to progress into a new area- or for power ups.